@@ -184,13 +184,9 @@ export default class extends Component {
   autoplayTimer = null
   loopJumpTimer = null
 
-  componentWillReceiveProps (nextProps) {
-    if (!nextProps.autoplay && this.autoplayTimer) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    if (!nextProps.autoplay && this.autoplayTimer){
       clearTimeout(this.autoplayTimer)
-      this.setState(this.initState(nextProps))
-    } else if(nextProps.autoplay){
-      this.autoplay(nextProps.autoplay);
-      this.loopJump();
     }
   }
 
@@ -203,9 +199,17 @@ export default class extends Component {
     this.loopJumpTimer && clearTimeout(this.loopJumpTimer)
   }
 
-  componentWillUpdate (nextProps, nextState) {
+  UNSAFE_componentWillUpdate(nextProps, nextState) {
     // If the index has changed, we notify the parent via the onIndexChanged callback
-    if (this.state.index !== nextState.index) this.props.onIndexChanged(nextState.index)
+    if (this.state.index !== nextState.index)
+      this.props.onIndexChanged(nextState.index)
+  }
+
+  componentDidUpdate(prevProps) {
+    // If autoplay props updated to true, autoplay immediately
+    if (this.props.autoplay && !prevProps.autoplay) {
+      this.autoplay()
+    }
   }
 
   initState (props) {
@@ -260,7 +264,7 @@ export default class extends Component {
 
     // only update the offset in state if needed, updating offset while swiping
     // causes some bad jumping / stuttering
-    if (!this.state.offset || width !== this.state.width || height !== this.state.height) {
+    if (!this.state.offset) {
       state.offset = offset
     }
     this.setState(state)
@@ -269,9 +273,37 @@ export default class extends Component {
   loopJump = () => {
     if (!this.state.loopJump) return
     const i = this.state.index + (this.props.loop ? 1 : 0)
-    const scrollView = this.refs.scrollView
-    this.loopJumpTimer = setTimeout(() => scrollView.setPageWithoutAnimation &&
-      scrollView.setPageWithoutAnimation(i), 50)
+    const scrollView = this.scrollView
+    this.loopJumpTimer = setTimeout(
+      () => {
+        if (scrollView.setPageWithoutAnimation) {
+          scrollView.setPageWithoutAnimation(i)
+        } else {
+          if (this.state.index === 0) {
+            scrollView.scrollTo(
+              this.props.horizontal === false
+                ? { x: 0, y: this.state.height, animated: false }
+                : { x: this.state.width, y: 0, animated: false }
+            )
+          } else if (this.state.index === this.state.total - 1) {
+            this.props.horizontal === false
+              ? this.scrollView.scrollTo({
+                  x: 0,
+                  y: this.state.height * this.state.total,
+                  animated: false
+                })
+              : this.scrollView.scrollTo({
+                  x: this.state.width * this.state.total,
+                  y: 0,
+                  animated: false
+                })
+          }
+        }
+      },
+      // Important Parameter
+      // ViewPager 50ms, ScrollView 300ms
+      scrollView.setPageWithoutAnimation ? 50 : 300
+    )
   }
 
   /**
@@ -327,10 +359,10 @@ export default class extends Component {
     this.updateIndex(e.nativeEvent.contentOffset, this.state.dir, () => {
       this.autoplay()
       this.loopJump()
-
-      // if `onMomentumScrollEnd` registered will be called here
-      this.props.onMomentumScrollEnd && this.props.onMomentumScrollEnd(e, this.fullState(), this)
     })
+    // if `onMomentumScrollEnd` registered will be called here
+    this.props.onMomentumScrollEnd &&
+      this.props.onMomentumScrollEnd(e, this.fullState(), this)
   }
 
   /*
@@ -358,6 +390,7 @@ export default class extends Component {
    */
   updateIndex = (offset, dir, cb) => {
     const state = this.state
+    // Android ScrollView will not scrollTo certain offset when props change
     let index = state.index
     const diff = offset[dir] - this.internals.offset[dir]
     const step = dir === 'x' ? state.width : state.height
@@ -653,9 +686,19 @@ export default class extends Component {
 
       pages = pages.map((page, i) => {
         if (loadMinimal) {
-          if (i >= (index + loopVal - loadMinimalSize) &&
-            i <= (index + loopVal + loadMinimalSize)) {
-            return <View style={pageStyle} key={i}>{children[page]}</View>
+          if (
+            (i >= index + loopVal - loadMinimalSize &&
+              i <= index + loopVal + loadMinimalSize) ||
+            // The real first swiper should be keep
+            (loop && i === 1) ||
+            // The real last swiper should be keep
+            (loop && i === total - 1)
+          ) {
+            return (
+              <View style={pageStyle} key={i}>
+                {children[page]}
+              </View>
+            )
           } else {
             return (
               <View style={pageStyleLoading} key={i}>
