@@ -1,13 +1,9 @@
 import React from 'react'
-import {
-  Text,
-  View
-} from 'react-native'
+import { Text, View } from 'react-native'
 import Swiper from 'react-native-swiper'
 
 var styles = {
-  wrapper: {
-  },
+  wrapper: {},
   slide1: {
     flex: 1,
     justifyContent: 'center',
@@ -33,14 +29,16 @@ var styles = {
   }
 }
 
-export default () => <Swiper style={styles.wrapper} showsButtons>
-  <View style={styles.slide1}>
-    <Text style={styles.text}>Hello Swiper</Text>
-  </View>
-  <View style={styles.slide2}>
-    <Text style={styles.text}>Beautiful</Text>
-  </View>
-  <View style={styles.slide3}>
-    <Text style={styles.text}>And simple</Text>
-  </View>
-</Swiper>
+export default () => (
+  <Swiper style={styles.wrapper} showsButtons loop={false}>
+    <View testID="Hello" style={styles.slide1}>
+      <Text style={styles.text}>Hello Swiper</Text>
+    </View>
+    <View testID="Beautiful" style={styles.slide2}>
+      <Text style={styles.text}>Beautiful</Text>
+    </View>
+    <View testID="Simple" style={styles.slide3}>
+      <Text style={styles.text}>And simple</Text>
+    </View>
+  </Swiper>
+)
