@@ -48,47 +48,62 @@ var styles = {
 
 const renderPagination = (index, total, context) => {
   return (
-    <View style={{
-      position: 'absolute',
-      justifyContent: 'center',
-      alignItems: 'center',
-      top: 25,
-      left: 0,
-      right: 0
-    }}>
-      <View style={{
-        borderRadius: 7,
-        backgroundColor: 'rgba(255,255,255,.15)',
-        padding: 3,
-        paddingHorizontal: 7
-      }}>
-        <Text style={{
-          color: '#fff',
-          fontSize: 14
-        }}>{index + 1} / {total}</Text>
+    <View
+      style={{
+        position: 'absolute',
+        justifyContent: 'center',
+        alignItems: 'center',
+        top: 25,
+        left: 0,
+        right: 0
+      }}
+    >
+      <View
+        style={{
+          borderRadius: 7,
+          backgroundColor: 'rgba(255,255,255,.15)',
+          padding: 3,
+          paddingHorizontal: 7
+        }}
+      >
+        <Text
+          style={{
+            color: '#fff',
+            fontSize: 14
+          }}
+        >
+          {index + 1} / {total}
+        </Text>
       </View>
     </View>
   )
 }
 
-const Viewer = props => <Swiper index={props.index} style={styles.wrapper} renderPagination={renderPagination}>
-  {
-    props.imgList.map((item, i) => <View key={i} style={styles.slide}>
-      <TouchableWithoutFeedback onPress={e => props.pressHandle()}>
-        <PhotoView
-          source={{uri: item}}
-          resizeMode='contain'
-          minimumZoomScale={0.5}
-          maximumZoomScale={3}
-          androidScaleType='center'
-          style={styles.photo} />
-      </TouchableWithoutFeedback>
-    </View>)
-  }
-</Swiper>
+const Viewer = props => (
+  <Swiper
+    index={props.index}
+    style={styles.wrapper}
+    renderPagination={renderPagination}
+  >
+    {props.imgList.map((item, i) => (
+      <View key={i} style={styles.slide}>
+        <TouchableWithoutFeedback onPress={e => props.pressHandle()}>
+          <PhotoView
+            source={{ uri: item }}
+            resizeMode="contain"
+            minimumZoomScale={0.5}
+            maximumZoomScale={3}
+            androidScaleType="center"
+            style={styles.photo}
+          />
+        </TouchableWithoutFeedback>
+      </View>
+    ))}
+  </Swiper>
+)
 
 export default class extends Component {
-  constructor (props) {
+  constructor(props) {
     super(props)
     this.state = {
       imgList: [
@@ -102,30 +117,35 @@ export default class extends Component {
     this.viewerPressHandle = this.viewerPressHandle.bind(this)
     this.thumbPressHandle = this.thumbPressHandle.bind(this)
   }
-  viewerPressHandle () {
+  viewerPressHandle() {
     this.setState({
       showViewer: false
     })
   }
-  thumbPressHandle (i) {
+  thumbPressHandle(i) {
     this.setState({
       showIndex: i,
       showViewer: true
     })
   }
-  render () {
-    return (<View style={{position: 'relative'}}>
-      {this.state.showViewer && <Viewer
-        index={this.state.showIndex}
-        pressHandle={this.viewerPressHandle}
-        imgList={this.state.imgList} />}
-      <View style={styles.thumbWrap}>
-        {
-          this.state.imgList.map((item, i) => <TouchableOpacity key={i} onPress={e => this.thumbPressHandle(i)}>
-            <Image style={styles.thumb} source={{uri: item}} />
-          </TouchableOpacity>)
-        }
+  render() {
+    return (
+      <View style={{ position: 'relative' }}>
+        {this.state.showViewer && (
+          <Viewer
+            index={this.state.showIndex}
+            pressHandle={this.viewerPressHandle}
+            imgList={this.state.imgList}
+          />
+        )}
+        <View style={styles.thumbWrap}>
+          {this.state.imgList.map((item, i) => (
+            <TouchableOpacity key={i} onPress={e => this.thumbPressHandle(i)}>
+              <Image style={styles.thumb} source={{ uri: item }} />
+            </TouchableOpacity>
+          ))}
+        </View>
       </View>
-    </View>)
+    )
   }
 }
