@@ -1,16 +1,10 @@
 import React, { Component } from 'react'
-import {
-  Text,
-  View,
-  Image,
-  Dimensions
-} from 'react-native'
+import { Text, View, Image, Dimensions } from 'react-native'
 import Swiper from 'react-native-swiper'
 const { width } = Dimensions.get('window')
 
 const styles = {
-  wrapper: {
-  },
+  wrapper: {},
   slide: {
     flex: 1,
     justifyContent: 'center',
@@ -47,26 +41,42 @@ const renderPagination = (index, total, context) => {
 }
 
 export default class extends Component {
-  render () {
+  render() {
     return (
-        <Swiper
-          style={styles.wrapper}
-          renderPagination={renderPagination}
-          loop={false}
+      <Swiper
+        style={styles.wrapper}
+        renderPagination={renderPagination}
+        loop={false}
+      >
+        <View
+          style={styles.slide}
+          title={
+            <Text numberOfLines={1}>Aussie tourist dies at Bali hotel</Text>
+          }
         >
-          <View style={styles.slide} title={<Text numberOfLines={1}>Aussie tourist dies at Bali hotel</Text>}>
-            <Image style={styles.image} source={require('./img/1.jpg')} />
-          </View>
-          <View style={styles.slide} title={<Text numberOfLines={1}>Big lie behind Nine’s new show</Text>}>
-            <Image style={styles.image} source={require('./img/2.jpg')} />
-          </View>
-          <View style={styles.slide} title={<Text numberOfLines={1}>Why Stone split from Garfield</Text>}>
-            <Image style={styles.image} source={require('./img/3.jpg')} />
-          </View>
-          <View style={styles.slide} title={<Text numberOfLines={1}>Learn from Kim K to land that job</Text>}>
-            <Image style={styles.image} source={require('./img/4.jpg')} />
-          </View>
-        </Swiper>
+          <Image style={styles.image} source={require('./img/1.jpg')} />
+        </View>
+        <View
+          style={styles.slide}
+          title={<Text numberOfLines={1}>Big lie behind Nine’s new show</Text>}
+        >
+          <Image style={styles.image} source={require('./img/2.jpg')} />
+        </View>
+        <View
+          style={styles.slide}
+          title={<Text numberOfLines={1}>Why Stone split from Garfield</Text>}
+        >
+          <Image style={styles.image} source={require('./img/3.jpg')} />
+        </View>
+        <View
+          style={styles.slide}
+          title={
+            <Text numberOfLines={1}>Learn from Kim K to land that job</Text>
+          }
+        >
+          <Image style={styles.image} source={require('./img/4.jpg')} />
+        </View>
+      </Swiper>
     )
   }
 }
